✨ feat(Finder):

+++ 21_位运算/二进制枚举与三进制枚举/枚举子集/enumerateBits.go
@@ -11,15 +11,15 @@ import (
 func main() {
 	time1 := time.Now()
 	for i := 0; i < 1e7; i++ {
-		EnumerateBits(i, func(bit int) {})
+		EnumerateBits(uint(i), func(bit int) {})
 	}
 	fmt.Println(time.Since(time1))
 }
 
 // 遍历每个为1的比特位
-func EnumerateBits(s int, f func(bit int)) {
+func EnumerateBits(s uint, f func(bit int)) {
 	for s > 0 {
-		i := bits.TrailingZeros(uint(s))
+		i := bits.TrailingZeros(s)
 		f(i)
 		s ^= 1 << i
 	}
